requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
pydantic>=2.10.0
python-dotenv>=1.0.0

# Vector database and embeddings
chromadb>=0.5.0
sentence-transformers>=2.5.0

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# LLM integration
openai>=1.50.0
anthropic>=0.25.0
ollama>=0.4.2

# LangChain and LangGraph for agentic RAG
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.2.0
langchain-core>=0.2.0

# AWS S3
boto3>=1.34.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Utilities
pandas>=2.2.0
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiohttp>=3.8.0
aiofiles>=23.1.0
scikit-learn>=1.3.0

# CORS
fastapi-cors>=0.0.6

# Environment and config
pydantic-settings>=2.1.0 